rename file name to have the same style
